fixed the spectate matches button that got messed up in merge
